refactor: decompose monolithic navbar into sub-components
